requirements.txt
timm==0.9.16
transformers==4.21.0
lmdb==1.4.1
Pillow==8.4.0
ftfy==6.2.0
joblib==1.3.2
opencv-python==4.7.0.72
pydantic
fastapi
uvicorn